realistic_samples/docshare/requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.5
bcrypt==4.0.1
PyJWT==2.4.0
requests==2.28.1
PyYAML==5.4.1
lxml==4.9.1
pycryptodome==3.17
redis==4.5.1
